reorganize sidebar files
