+++ apps/meteor/client/sidebar/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Sidebar';
+export { default } from './SidebarRegion';
