+++ apps/meteor/client/views/modal/ModalRegion.tsx
@@ -1,10 +1,11 @@
 import { useModal, useCurrentModal } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
-import React, { useCallback } from 'react';
-import { FocusScope } from 'react-aria';
+import React, { lazy, useCallback } from 'react';
 
 import ModalBackdrop from '../../components/modal/ModalBackdrop';
 import ModalPortal from '../../components/modal/ModalPortal';
+
+const FocusScope = lazy(() => import('react-aria').then((module) => ({ default: module.FocusScope })));
 
 const ModalRegion = (): ReactElement | null => {
 	const currentModal = useCurrentModal();
