+++ apps/meteor/client/views/room/Header/Header.tsx
@@ -3,13 +3,14 @@ import { isVoipRoom } from '@rocket.chat/core-typings';
 import { HeaderToolbox } from '@rocket.chat/ui-client';
 import { useLayout } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
-import React, { memo, useMemo } from 'react';
+import React, { lazy, memo, useMemo } from 'react';
 
 import BurgerMenu from '../../../components/BurgerMenu';
-import DirectRoomHeader from './DirectRoomHeader';
-import OmnichannelRoomHeader from './Omnichannel/OmnichannelRoomHeader';
-import VoipRoomHeader from './Omnichannel/VoipRoomHeader';
-import RoomHeader from './RoomHeader';
+
+const DirectRoomHeader = lazy(() => import('./DirectRoomHeader'));
+const OmnichannelRoomHeader = lazy(() => import('./Omnichannel/OmnichannelRoomHeader'));
+const VoipRoomHeader = lazy(() => import('./Omnichannel/VoipRoomHeader'));
+const RoomHeader = lazy(() => import('./RoomHeader'));
 
 type HeaderProps<T> = {
 	room: T;
